+++ tracing/openai/src/jvmMain/kotlin/org/jetbrains/ai/tracy/openai/adapters/handlers/ResponsesOpenAIApiEndpointHandler.kt
@@ -422,7 +422,6 @@ internal class ResponsesOpenAIApiEndpointHandler(
     private val mappedResponseAttributes: List<String> = listOf(
         // parsed by `OpenAIApiUtils.setCommonResponseAttributes`
         "id",
-        "object",
         "model",
 
         "output",